refactor(errors): move them in models folder
