+++ flutter_reach_five/lib/flutter_reach_five.dart
@@ -1,10 +1,10 @@
 library flutter_reach_five;
 
-export 'errors.dart';
 export 'models/address.dart';
 export 'models/auth_token.dart';
 export 'models/consent.dart';
 export 'models/emails.dart';
+export 'models/errors.dart';
 export 'models/login_summary.dart';
 export 'models/open_id_user.dart';
 export 'models/profile.dart';
